# src/swaggerOutput.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "v1.0.0",
    "title": "TradeMark",
    "description": "TradeMark REST API Documentation"
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/",
      "description": "Local Host"
    }
  ],
  "paths": {
...
    "/withdrawal/all": {
      "get": {
        "tags": ["User Withdrawals"],
        "description": "",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WithdrawalsResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/WithdrawalsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidToken"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidToken"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/admin/withdrawals/all": {
      "get": {
        "tags": ["Admin - Withdrawals"],
        "description": "",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WithdrawalsResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/WithdrawalsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidToken"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidToken"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotExists"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/NotExists"
                }
              }
            }
          },
          "406": {
            "description": "Not Acceptable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidID"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidID"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/admin/withdrawals/user/{uid}": {
      "get": {
        "tags": ["Admin - Withdrawals"],
        "description": "",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WithdrawalsResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/WithdrawalsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidToken"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidToken"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotExists"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/NotExists"
                }
              }
            }
          },
          "406": {
            "description": "Not Acceptable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidID"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidID"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/admin/withdrawals/{id}/approve": {
      "patch": {
        "tags": ["Admin - Withdrawals"],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "406": {
            "description": "Not Acceptable"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/admin/withdrawals/{id}/processing": {
      "patch": {
        "tags": ["Admin - Withdrawals"],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "406": {
            "description": "Not Acceptable"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/admin/withdrawals/{id}/decline": {
      "patch": {
        "tags": ["Admin - Withdrawals"],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "406": {
            "description": "Not Acceptable"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "WithdrawalsResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "example": "string"
                },
                "uid": {
                  "type": "string",
                  "example": "string"
                },
                "amount": {
                  "type": "string",
                  "example": "number"
                },
                "to": {
                  "type": "string",
                  "example": "string"
                },
                "method": {
                  "type": "string",
                  "example": "string"
                },
                "status": {
                  "type": "string",
                  "example": "string [pending | approved | declined]"
                },
                "createdAt": {
                  "type": "string",
                  "example": "number"
                },
                "updatedAt": {
                  "type": "string",
                  "example": "number"
                }
              }
            }
          }
        },
        "xml": {
          "name": "WithdrawalsResponse"
        }
      },
...
      "InvalidToken": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Invalid Token."
          }
        },
        "xml": {
          "name": "InvalidToken"
        }
      },
...
      "NotExists": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Does not exists."
          }
        },
        "xml": {
          "name": "NotExists"
        }
      },
      "InvalidID": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Invalid ID."
          }
        },
        "xml": {
          "name": "InvalidID"
        }
      },
...
    }
  },
...
}